Hoist ProfileContext.Provider up to the [id].tsx page, for player view purposes
