+++ components/Player/Profile.tsx
@@ -1,27 +1,16 @@
-import { Absence, AbsenceType, ProfileFieldKey } from "@prisma/client";
-import React, { useCallback, useContext, useEffect, useState } from "react";
+import { AbsenceType, ProfileFieldKey } from "@prisma/client";
+import React, { useContext, useState } from "react";
 import Icon from "components/Icon";
 import {
-  IAbsence,
   IPlayer,
-  IProfileField,
   ProfileCategory,
   ProfileCategoryIcons,
   ProfileFieldsByCategory,
-  ProfileFieldValueDeserializedTypes,
-  ProfileFieldValues,
 } from "interfaces";
-import { useRouter } from "next/router";
-import { UpdateOneAbsenceDTO } from "pages/api/absences/update";
-import { serializeProfileFieldValue } from "utils/buildUserProfile";
-import isAbsence from "utils/isAbsence";
 import AbsenceTable from "./AbsenceTable";
 import ProfileFieldCell from "./ProfileFieldCell";
 import NotesTable from "./NotesTable";
-import ProfileContext, {
-  ProfileContextType,
-  useProfileContext,
-} from "./ProfileContext";
+import ProfileContext from "./ProfileContext";
 import ProfileSection, { Props as ProfileSectionProps } from "./ProfileSection";
 
 type ProfileContentsProps<T extends ProfileCategory> = {
@@ -143,112 +132,6 @@ const Profile: React.FunctionComponent<Props> = ({ player }: Props) => {
   const [selectedCategory, setSelectedCategory] = useState(
     ProfileCategory.Overview
   );
-  const [state, dispatch] = useProfileContext();
-  const router = useRouter();
-
-  const refreshProfile = useCallback(() => {
-    router.replace(router.asPath);
-  }, [router]);
-
-  const createField: ProfileContextType["createField"] = useCallback(
-    async function createField(fieldKey, draft, userId): Promise<void> {
-      const serializedValue = JSON.stringify(
-        fieldKey === "absence"
-          ? {
-              absences: [draft as Partial<Absence>],
-              playerId: userId,
-            }
-          : {
-              playerId: userId,
-              fields: [
-                {
-                  key: fieldKey,
-                  value: serializeProfileFieldValue(
-                    draft as ProfileFieldValueDeserializedTypes[ProfileFieldValues[ProfileFieldKey]],
-                    fieldKey
-                  ),
-                },
-              ],
-            }
-      );
-
-      const response = await fetch(
-        fieldKey === "absence" ? "/api/absences" : "/api/profileFields",
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          credentials: "include",
-          body: serializedValue,
-        }
-      );
-      if (!response.ok) {
-        throw await response.json();
-      }
-      refreshProfile();
-    },
-    [refreshProfile]
-  );
-
-  const updateField: ProfileContextType["updateField"] = useCallback(
-    async function updateField(field: IProfileField | IAbsence): Promise<void> {
-      if (!field.draft) {
-        return;
-      }
-      const serializedValue = isAbsence(field)
-        ? JSON.stringify({
-            date: field.draft?.date,
-            description: field.draft?.description,
-            reason: field.draft?.reason,
-            type: field.draft?.type,
-            userId: field.draft?.userId,
-          } as UpdateOneAbsenceDTO)
-        : JSON.stringify({
-            value: serializeProfileFieldValue(field.draft, field.key),
-          });
-      const response = await fetch(
-        isAbsence(field)
-          ? `/api/absences/${field.id}`
-          : `/api/profileFields/${field.id}`,
-        {
-          method: "PATCH",
-          headers: { "Content-Type": "application/json" },
-          credentials: "include",
-          body: serializedValue,
-        }
-      );
-      if (!response.ok) {
-        throw await response.json();
-      }
-      refreshProfile();
-    },
-    [refreshProfile]
-  );
-
-  const deleteField: ProfileContextType["deleteField"] = useCallback(
-    async function deleteField(
-      fieldKey: ProfileFieldKey | "absence",
-      id: number
-    ): Promise<void> {
-      const response = await fetch(
-        fieldKey === "absence"
-          ? `/api/absences/${id}`
-          : `/api/profileFields/${id}`,
-        {
-          method: "DELETE",
-          credentials: "include",
-        }
-      );
-      if (!response.ok) {
-        throw await response.json();
-      }
-      refreshProfile();
-    },
-    [refreshProfile]
-  );
-
-  useEffect(() => {
-    dispatch({ type: "SET_PLAYER", player });
-  }, [player, dispatch]);
 
   return (
     <div className="pb-24">
@@ -284,13 +167,9 @@ const Profile: React.FunctionComponent<Props> = ({ player }: Props) => {
           ))}
       </div>
       <hr className="my-10" />
-      <ProfileContext.Provider
-        value={{ state, dispatch, createField, updateField, deleteField }}
-      >
-        <h1 className="mb-10 text-2xl font-semibold">{selectedCategory}</h1>
-        <hr />
-        <ProfileContents category={selectedCategory} />
-      </ProfileContext.Provider>
+      <h1 className="mb-10 text-2xl font-semibold">{selectedCategory}</h1>
+      <hr />
+      <ProfileContents category={selectedCategory} />
     </div>
   );
 };
